day02 ok

+++ day03/ex01/ScavTrap.cpp
@@ -6,7 +6,7 @@
 /*   By: dboyer <dboyer@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/27 14:25:40 by dboyer            #+#    #+#             */
-/*   Updated: 2020/12/26 11:38:34 by dboyer           ###   ########.fr       */
+/*   Updated: 2020/12/26 11:41:15 by dboyer           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -170,7 +170,7 @@ void ScavTrap::takeDamage(unsigned int amount)
 void ScavTrap::beRepaired(unsigned int amount)
 {
 	const int newHp = amount + this->_hitPoint;
-	if (amount > this->_maxHitPoint)
+	if ((int)amount > this->_maxHitPoint)
 	{
 		this->_hitPoint = this->_maxHitPoint;
 	}
